pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "code-review-env"
version = "1.0.0"
description = "OpenEnv environment for AI code review benchmarking"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    "openenv-core>=0.1.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.0.0",
    "scikit-learn>=1.3.0",
    "sentence-transformers>=2.2.0",
    "numpy>=1.24.0",
    "huggingface-hub>=0.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "httpx>=0.25.0",
]
baseline = [
    "openai>=1.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["server*", "data*"]

[tool.setuptools.package-data]
data = ["**/*.json"]
